Using long but can't update recommendation

+++ src/main/java/com/hsbc/roboadvisor/model/Portfolio/Portfolio.java
@@ -7,7 +7,7 @@ public class Portfolio implements Serializable
 {
     private String customerId;
 
-    private Integer id;
+    private Long id;
 
     private List<Holding> holdings;
 
@@ -15,7 +15,7 @@ public class Portfolio implements Serializable
         //empty constructor
     }
 
-    public Portfolio(String customerId, Integer id, List<Holding> holdings) {
+    public Portfolio(String customerId, Long id, List<Holding> holdings) {
         this.customerId = customerId;
         this.id = id;
         this.holdings = holdings;
@@ -29,11 +29,11 @@ public class Portfolio implements Serializable
         this.customerId = customerId;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
