+++ src/main/java/com/hsbc/roboadvisor/payload/TransactionRequest.java
@@ -1,17 +1,17 @@
 package com.hsbc.roboadvisor.payload;
 
 
+import java.util.List;
+import javax.validation.constraints.NotNull;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.hsbc.roboadvisor.model.Recommendation.Transaction;
-
-import javax.validation.constraints.NotNull;
-import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class TransactionRequest {
 
     @NotNull
-    private Integer portfolioId;
+    private Long portfolioId;
 
     @NotNull
     private List<Transaction> instructions;
@@ -19,16 +19,16 @@ public class TransactionRequest {
     public TransactionRequest() {
     }
 
-    public TransactionRequest(@NotNull Integer portfolioId, @NotNull List<Transaction> instructions) {
+    public TransactionRequest(@NotNull Long portfolioId, @NotNull List<Transaction> instructions) {
         this.portfolioId = portfolioId;
         this.instructions = instructions;
     }
 
-    public Integer getPortfolioId() {
+    public Long getPortfolioId() {
         return portfolioId;
     }
 
-    public void setPortfolioId(Integer portfolioId) {
+    public void setPortfolioId(Long portfolioId) {
         this.portfolioId = portfolioId;
     }
