+++ src/main/java/com/hsbc/roboadvisor/service/PortfolioRepositoryService.java
@@ -19,15 +19,15 @@ public class PortfolioRepositoryService
     @Autowired
     private PortfolioRepository portfolioRepository;
 
-    public PortfolioPreference findPreferenceByPortfolioId(Integer portfolioId) {
+    public PortfolioPreference findPreferenceByPortfolioId(Long portfolioId) {
         return this.portfolioRepository.findByPortfolioId(portfolioId);
     }
 
-    public Boolean preferenceExistsByPortfolioId(Integer portfolioId) {
+    public Boolean preferenceExistsByPortfolioId(Long portfolioId) {
         return this.portfolioRepository.existsByPortfolioId(portfolioId);
     }
 
-    public PortfolioPreference savePreference(Integer portfolioId, PortfolioRequest portfolioRequest) {
+    public PortfolioPreference savePreference(Long portfolioId, PortfolioRequest portfolioRequest) {
         totalPercentIs100OrFail(portfolioRequest.getAllocations());
 
         PortfolioPreference portfolio = new PortfolioPreference(portfolioId, portfolioRequest.getDeviation(),
@@ -37,14 +37,14 @@ public class PortfolioRepositoryService
     }
 
 
-    public PortfolioPreference updateDeviationByPortfolioId(Integer portfolioId, DeviationRequest deviationRequest) {
+    public PortfolioPreference updateDeviationByPortfolioId(Long portfolioId, DeviationRequest deviationRequest) {
         PortfolioPreference portfolio = this.portfolioRepository.findByPortfolioId(portfolioId);
         portfolio.setDeviation(deviationRequest.getDeviation());
 
         return this.portfolioRepository.save(portfolio);
     }
 
-    public PortfolioPreference updateAllocationsByPortfolioId(Integer portfolioId, List<Allocation> allocationList) {
+    public PortfolioPreference updateAllocationsByPortfolioId(Long portfolioId, List<Allocation> allocationList) {
         totalPercentIs100OrFail(allocationList);
 
         PortfolioPreference portfolio = this.portfolioRepository.findByPortfolioId(portfolioId);
